[#109] Refactor: 패키지 구조 변경
 - Jay와 얘기해서 패키지 구조를 변경하였습니다.
