+++ BE/src/main/java/com/codesquad/team1/signup/common/config/WebConfig.java
@@ -1,4 +1,4 @@
-package com.codesquad.team1.signup.config;
+package com.codesquad.team1.signup.common.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
